backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pymediainfo>=6.1.0
librosa>=0.10.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
scipy>=1.11.0
soundfile>=0.12.0
sqlalchemy>=2.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
transformers>=4.40.0
torch>=2.0.0
huggingface_hub>=0.23.0
streamlit>=1.35.0
pandas>=2.0.0
